+++ apps/analytics/components/Table/constants.tsx
@@ -1,4 +1,4 @@
-import { Pair, Token } from '@sushiswap/graph-client/.graphclient'
+import { Pair, Token } from '@sushiswap/graph-client'
 import { ColumnDef } from '@tanstack/react-table'
 import React from 'react'