+++ apps/earn/components/PoolPositionStakedProvider.tsx
@@ -1,5 +1,5 @@
 import { Amount, Currency, Token } from '@sushiswap/currency'
-import { Pair } from '@sushiswap/graph-client/.graphclient'
+import { Pair } from '@sushiswap/graph-client'
 import { Chef, useMasterChef } from '@sushiswap/wagmi'
 import { createContext, FC, ReactNode, useContext, useMemo } from 'react'