+++ packages/cli/actions/index.ts
@@ -1,4 +1,0 @@
-export * from './bar'
-export * from './chef'
-export * from './maker'
-export * from './serve'