+++ packages/config/graph/src/index.ts
@@ -31,7 +31,7 @@ export const PENDING_KAVA_HOST = 'pvt.graph.kava.io/subgraphs/id'
 export const METIS_HOST = 'andromeda.thegraph.metis.io/subgraphs/name'
 export const PENDING_METIS_HOST = 'andromeda.thegraph.metis.io/subgraphs/id'
 
-export const CHAIN_NAME: Record<number | string, string> = {
+export const CHAIN_NAME: Record<number, string> = {
   [ChainId.ARBITRUM]: 'Arbitrum',
   [ChainId.AVALANCHE]: 'Avalanche',
   [ChainId.BSC]: 'Bsc',
@@ -57,7 +57,7 @@ export const CHAIN_NAME: Record<number | string, string> = {
   [ChainId.BOBA_AVAX]: 'Boba Avax',
 }
 
-export const SUBGRAPH_HOST: Record<number | string, string> = {
+export const SUBGRAPH_HOST: Record<number, string> = {
   [ChainId.ARBITRUM]: GRAPH_HOST,
   [ChainId.ARBITRUM_NOVA]: 'arbitrum-nova-subgraph.sushi.com/subgraphs/name',
   [ChainId.AVALANCHE]: GRAPH_HOST,
@@ -76,9 +76,12 @@ export const SUBGRAPH_HOST: Record<number | string, string> = {
   [ChainId.OPTIMISM]: 'api.thegraph.com/subgraphs/name',
   [ChainId.POLYGON]: GRAPH_HOST,
   [ChainId.BOBA]: GRAPH_HOST,
+  [ChainId.OKEX]: '',
+  [ChainId.HECO]: '',
+  [ChainId.KOVAN]: '',
 }
 
-export const BENTOBOX_SUBGRAPH_NAME: Record<number | string, string> = {
+export const BENTOBOX_SUBGRAPH_NAME = {
   [ChainId.ETHEREUM]: 'sushiswap/bentobox-ethereum',
   [ChainId.POLYGON]: 'sushiswap/bentobox-polygon',
   [ChainId.AVALANCHE]: 'sushiswap/bentobox-avalanche',
@@ -92,9 +95,9 @@ export const BENTOBOX_SUBGRAPH_NAME: Record<number | string, string> = {
   [ChainId.OPTIMISM]: 'sushiswap/bentobox-optimism',
   [ChainId.HARMONY]: 'sushiswap/bentobox-harmony',
   [ChainId.KAVA]: 'sushiswap/bentobox-kava',
-}
+} as const
 
-export const BLOCKS_SUBGRAPH_NAME: Record<number | string, string> = {
+export const BLOCKS_SUBGRAPH_NAME: Record<number, string> = {
   [ChainId.ETHEREUM]: 'blocklytics/ethereum-blocks',
   [ChainId.GNOSIS]: 'matthewlilley/xdai-blocks',
   [ChainId.POLYGON]: 'matthewlilley/polygon-blocks',
@@ -117,7 +120,7 @@ export const BLOCKS_SUBGRAPH_NAME: Record<number | string, string> = {
   [ChainId.BOBA]: 'sushiswap/blocks-boba',
 }
 
-export const EXCHANGE_SUBGRAPH_NAME: Record<number | string, string> = {
+export const EXCHANGE_SUBGRAPH_NAME: Record<number, string> = {
   [ChainId.AVALANCHE]: 'sushiswap/exchange-avalanche',
   [ChainId.ARBITRUM]: 'sushiswap/exchange-arbitrum-backup',
   [ChainId.BSC]: 'sushiswap/bsc-exchange',
@@ -135,7 +138,7 @@ export const EXCHANGE_SUBGRAPH_NAME: Record<number | string, string> = {
   [ChainId.BOBA]: 'sushi-0m/sushiswap-boba',
 }
 
-export const SUSHISWAP_SUBGRAPH_NAME: Record<number | string, string> = {
+export const SUSHISWAP_SUBGRAPH_NAME = {
   [ChainId.ETHEREUM]: 'subgraph-qa/sushiswap-ethereum',
   [ChainId.AVALANCHE]: 'sushi-0m/sushiswap-avalanche',
   [ChainId.ARBITRUM]: 'sushi-0m/sushiswap-arbitrum',
@@ -150,7 +153,7 @@ export const SUSHISWAP_SUBGRAPH_NAME: Record<number | string, string> = {
   [ChainId.ARBITRUM_NOVA]: 'sushi-0m/sushiswap-arbitrum-nova',
   [ChainId.BOBA]: 'sushi-0m/sushiswap-boba',
   [ChainId.POLYGON]: 'subgraph-qa/sushiswap-polygon',
-}
+} as const
 
 export const TRIDENT_SUBGRAPH_NAME = {
   [ChainId.POLYGON]: 'sushi-qa/trident-polygon',
@@ -166,7 +169,7 @@ export const TRIDENT_SUBGRAPH_START_BLOCK: Record<keyof typeof TRIDENT_SUBGRAPH_
   [ChainId.METIS]: 3030678,
 }
 
-export const MINICHEF_SUBGRAPH_NAME: Record<number | string, string> = {
+export const MINICHEF_SUBGRAPH_NAME = {
   [ChainId.POLYGON]: 'jiro-ono/minichef-staging-updates', // new trident subgraph not synced yet
   [ChainId.GNOSIS]: 'jiro-ono/gnosis-minichef-staging',
   // [ChainId.HARMONY]: 'sushiswap/harmony-minichef', // subgraph broken
@@ -179,11 +182,12 @@ export const MINICHEF_SUBGRAPH_NAME: Record<number | string, string> = {
   [ChainId.KAVA]: 'sushiswap/kava-minichef', //block subgraph not synced yet
   [ChainId.METIS]: 'sushiswap/metis-minichef',
   [ChainId.BOBA]: 'sushiswap/minichef-boba',
-}
+} as const
 
-export const MASTERCHEF_V1_SUBGRAPH_NAME = 'jiro-ono/masterchef-staging'
-export const MASTERCHEF_V2_SUBGRAPH_NAME = 'sushiswap/master-chefv2'
-export const FURO_SUBGRAPH_NAME: Record<string | number, string> = {
+export const MASTERCHEF_V1_SUBGRAPH_NAME = 'jiro-ono/masterchef-staging' as const
+export const MASTERCHEF_V2_SUBGRAPH_NAME = 'sushiswap/master-chefv2' as const
+
+export const FURO_SUBGRAPH_NAME: Record<string, string> = {
   [ChainId.ETHEREUM]: 'sushi-subgraphs/furo-ethereum',
   [ChainId.GÖRLI]: 'sushi-subgraphs/furo-goerli',
   [ChainId.ARBITRUM]: 'sushi-subgraphs/furo-arbitrum',
@@ -198,7 +202,7 @@ export const FURO_SUBGRAPH_NAME: Record<string | number, string> = {
   [ChainId.POLYGON]: 'sushi-subgraphs/furo-polygon',
 }
 
-export const KASHI_SUBGRAPH_NAME: Record<number | string, string> = {
+export const KASHI_SUBGRAPH_NAME: Record<number, string> = {
   [ChainId.ARBITRUM]: 'sushiswap/kashi-arbitrum',
 }