+++ packages/config/graph/CHANGELOG.md
@@ -1,8 +1,11 @@
 # @sushiswap/graph-config
 
-## 0.0.1
+## 0.1.0
+
+### Minor Changes
+
+- b3761c7c: initial release
 
 ### Patch Changes
 
-- Updated dependencies [a38f2e6a]
-  - @sushiswap/chain@0.1.0
+- @sushiswap/chain@0.1.0