+++ packages/tines/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @sushiswap/tines
 
+## 1.0.3
+
+### Patch Changes
+
+- Updated dependencies [e0b880fe]
+  - @sushiswap/currency@0.0.3
+
 ## 1.0.2
 
 ### Patch Changes